api_gateway/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
line-bot-sdk==3.4.0
redis==5.0.0
pydantic==2.4.2
httpx==0.25.0 